+++ benchmark/write.js
@@ -1,0 +1,4 @@
+/* global Promise */
+'use strict'
+var Benchmark = require('benchmark');
+
