+++ index.js
@@ -1,1 +1,9 @@
-module.exports = require('./lib/connection.js');
+// module.exports = require('./lib/connection.js');
+let T = require('./lib/connection.js');
+let conn = new T({port: 3313});
+conn.connect()
+    .then(()=>{console.log('connected 1')})
+    // .catch((e)=>{console.log(e)});
+// conn.connect()
+//     .then(()=>{console.log('connected')})
+//     .catch((e)=>{console.log(e)});
