+++ lib/const.js
