+++ lib/denque.js
