+++ lib/event-handler.js
@@ -1,0 +1,19 @@
+
+
+exports.dataHandler = function(self){
+    return function(data){
+                
+    };
+};
+
+exports.errorHandler = function(self){
+    return function(error){
+
+    };
+};
+
+exports.closeHandler = function(self){
+    return function(){
+
+    };
+};
