+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "lint": "eslint ./lib ./test",
-    "test": "mocha -t 5000"
+    "test": "mocha -t 5000",
+    "benchmark": "node benchmark/read"
   },
   "repository": {
     "type": "git",
@@ -25,6 +26,9 @@
   "homepage": "https://github.com/KlonD90/node-tarantool-driver",
   "dependencies": {
     "bitwise-xor": "0.0.0",
+    "debug": "^2.6.8",
+    "ioredis": "^3.1.1",
+    "lodash": "^4.17.4",
     "msgpack-lite": "^0.1.20",
     "object-assign": "^4.1.0",
     "underscore": "^1.8.3"
